% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioFuelData.R
\name{getBioFuelData}
\alias{getBioFuelData}
\title{Get Bio Fuel Data}
\usage{
  getBioFuelData(dataContext)
}
\arguments{
  \item{dataContext}{The data context}
}
\value{
  Codes
}
\description{
  Get Bio Fuel Data
}

