% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryCode.R
\name{getCountryCode}
\alias{getCountryCode}
\title{Get Country Code}
\usage{
getCountryCode(usdaCountryName)
}
\arguments{
\item{usdaCountryName}{The USDA country code of the country.}
}
\value{
Codes
}
\description{
Get Country Code
}

