% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryCodeSUA.R
\name{getCountryCodeSUA}
\alias{getCountryCodeSUA}
\title{Get Country Code SUA}
\usage{
getCountryCodeSUA(country_name)
}
\arguments{
\item{country_name}{The name of the country.}
}
\value{
Codes
}
\description{
Get Country Code SUA
}

