% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPCTreeItem.R
\name{getCPCTreeItem}
\alias{getCPCTreeItem}
\title{Get CPC Item Tree}
\usage{
getCPCTreeItem(itemVar = "measuredItemCPC")
}
\arguments{
\item{itemVar}{The name of the "item" dimension.}
}
\value{
Codes
}
\description{
Get CPC Item Tree
}

