% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndustrialData.R
\name{getIndustrialData}
\alias{getIndustrialData}
\title{Get Industrial Data}
\usage{
getIndustrialData(yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the industrial data.
}
\description{
This function pulls the industrial data from the SWS
}

