% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItemCommSUA.R
\name{getItemCommSUA}
\alias{getItemCommSUA}
\title{Get Item Comm SUA}
\usage{
getItemCommSUA(Measured_Item_Psd)
}
\arguments{
\item{Measured_Item_Psd}{The USDA item code .}
}
\value{
Codes
}
\description{
Get Item Comm SUA
}

