% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitIndustrialData.R
\name{splitIndustrialData}
\alias{splitIndustrialData}
\title{Split Industrial Data}
\usage{
splitIndustrialData(usdaData)
}
\arguments{
\item{usdaData}{The dataset from USDA.}
}
\description{
This function splits the industrial data from USDA for member States
of European Union.
}

