% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioFuelData.R
\name{getBioFuelData}
\alias{getBioFuelData}
\title{Get Bio Fuel Data}
\usage{
getBioFuelData(areaVar = "geographicAreaM49", elementVar = "induseElement",
  itemVar = "measuredItemCPC", yearVar = "timePointYears")
}
\arguments{
\item{areaVar}{The name of the "geographic area" dimension.}

\item{elementVar}{The name of the "element" dimension.}

\item{itemVar}{The name of the "item" dimension.}

\item{yearVar}{The name of the "year" dimension.}
}
\value{
Codes
}
\description{
Get Bio Fuel Data
}

