% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCPCTreeItem.R
\name{getCPCTreeItem}
\alias{getCPCTreeItem}
\title{Get CPC Item Tree}
\usage{
getCPCTreeItem(dataContext)
}
\arguments{
\item{dataContext}{The data context}
}
\value{
Codes
}
\description{
Get CPC Item Tree
}

