% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getItemCommSUA.R
\name{getItemCommSUA}
\alias{getItemCommSUA}
\title{Get Item Commodity of SUA}
\usage{
  getItemCommSUA(dataContext)
}
\arguments{
  \item{dataContext}{The data context}
}
\value{
  Codes
}
\description{
  Get Item Commodity of SUA
}